requirements.txt
fastapi
uvicorn
openai==2.8.0
python-dotenv
pydantic
slowapi